+++ components/pageComponents/Login.js
@@ -12,8 +12,6 @@ import { useNavigation } from "@react-navigation/native";
 import {
   getAuth,
   signInWithEmailAndPassword,
-  onAuthStateChanged,
-  signOut
 } from "firebase/auth";
 
 function Login() {
@@ -49,19 +47,6 @@ function Login() {
   //   }
   // });
 
-  // SIGN OUT FUNCTIONALITY
-  function testSignOut() {
-    signOut(auth)
-      .then(() => {
-        console.log("SIGNED OUT");
-        console.log(auth.currentUser);
-      
-      })
-      .catch((error) => {
-        console.log("ERROR");
-      });
-  }
-
   return (
     <View style={styles.container}>
       <Image
@@ -96,16 +81,6 @@ function Login() {
           }) => (
             <>
               <View style={styles.editFormContainer}>
-              <PrimaryButton
-                label="LogOut"
-                btnWidth={{
-                  width: 182,
-                  right: 50,
-                  bottom: -110,
-                  position: "absolute",
-                }}
-                onPress={() => testSignOut()}
-              />
                 <View style={styles.form}>
                   <Text
                     style={{
